feat: refactor waiting room arguments and repository implementation
