+++ src/main/java/com/icebreaker/be/infra/persistence/redis/waitingroom/JoinRoomArgs.java
@@ -1,6 +1,7 @@
-package com.icebreaker.be.infra.persistence.redis;
+package com.icebreaker.be.infra.persistence.redis.waitingroom;
 
 import com.icebreaker.be.domain.room.vo.WaitingRoomParticipant;
+import com.icebreaker.be.infra.persistence.redis.RedisArgs;
 import java.time.ZoneOffset;
 
 public record JoinRoomArgs(
